refactor(js): unify nav-helpers, toc-scroll-spy, infinite-scroll into scroll-manager
Replace 3 separate JS files with a single scroll-manager.js that handles:
- Back-to-top button visibility and keyboard-nav detection (was nav-helpers.js)
- TOC scroll-spy with scrollend-based highlighting (was toc-scroll-spy.js)
- Infinite scroll trigger detection and observer (was infinite-scroll.js)
- Scroll position save/restore with manual scrollRestoration

Bug fixes:
- b2: TOC not updating on browser back/forward between hash locations
  (finishNavigation now calls updateTocHighlight explicitly)
- b3: Wrong scroll position on back-nav with TOC hash mismatch
  (scroll key uses pathname+search only, no hash)

Architecture:
- Single 'navigating' boolean gates all scroll work
- onScroll: saves position, updates button, checks infinite scroll
- onScrollEnd: updates TOC highlight (scrollend event + debounce fallback)
- finishNavigation: called from markScriptsReady, clears gate + runs onScrollEnd
- scrollToHash: handles same-page hash popstate navigation

Delete old files: nav-helpers.js, toc-scroll-spy.js, infinite-scroll.js
Delete old tests: nav-helpers.test.js, toc-scroll-spy.test.js, infinite-scroll.test.js
Add unified test: scroll-manager.test.js (55 tests)
Update all references in docs, AGENTS.md, comments, E2E helpers
