+++ src/TechHub.Web/Components/App.razor
@@ -185,9 +185,9 @@
         </script>
     }
 
-    <!-- Navigation Helpers - Back to Top and Back to Previous -->
+    <!-- Scroll Manager - Navigation, scroll position, TOC scroll spy, infinite scroll -->
     <!-- Uses @Assets for fingerprinted URL (cache busting). See Configuration/JsFiles.cs -->
-    <script type="module" src="@Assets["js/nav-helpers.js"]"></script>
+    <script type="module" src="@Assets["js/scroll-manager.js"]"></script>
 
     <!-- Mobile Navigation - Scroll lock and keyboard handling for hamburger menu -->
     <!-- Uses @Assets for fingerprinted URL (cache busting). See Configuration/JsFiles.cs -->
