+++ src/TechHub.Web/wwwroot/TechHub.Web.lib.module.js
@@ -55,8 +55,8 @@ export function afterWebStarted(blazor) {
     // browser's async scroll restoration races with our synchronous scroll restore
     // (triggered by markScriptsReady), randomly clobbering the restored position with 0.
     // We take full ownership of scroll restoration:
-    //   - Forward navigation → nav-helpers.js resetPagePosition() scrolls to top
-    //   - Back/forward navigation → nav-helpers.js restoreScrollPosition() called via
+    //   - Forward navigation → scroll-manager.js resetPagePosition() scrolls to top
+    //   - Back/forward navigation → scroll-manager.js restoreScrollPosition() called via
     //     markScriptsReady after all rendering is complete
     if ('scrollRestoration' in history) {
         history.scrollRestoration = 'manual';
