--- src/TechHub.Web/wwwroot/js/toc-scroll-spy.js
@@ -1,482 +1,0 @@
-/**
- * Table of Contents Scroll Spy
- * 
- * Automatically highlights the TOC link for the currently visible heading.
- * Uses scrollend event for clean, reliable detection when scrolling stops.
- * 
- * CRITICAL: Uses history.replaceState() instead of pushState() to update URL hash.
- * This prevents polluting browser history with scroll positions - only actual
- * TOC link clicks create history entries, enabling clean back button navigation.
- * 
- * @see {@link /workspaces/techhub/src/TechHub.Web/AGENTS.md} for integration patterns
- */
-
-// Hardcoded constants - must match CSS scroll-margin-top value
-const STICKY_HEADER_OFFSET = 106; // px - height of sticky header
-const DETECTION_LINE_PERCENT = 0.30; // 30% from top of content area
-
-export class TocScrollSpy {
-    constructor(tocElement, contentElement) {
-        this.tocElement = tocElement;
-        this.contentElement = contentElement;
-        this.headings = [];
-        this.tocLinks = new Map(); // headingId -> tocLink element
-        this.headingElements = new Map(); // headingId -> heading element (cached)
-        this.h2Items = null; // Cached TOC h2 items (.toc-depth-0)
-        this.currentActiveId = null;
-        this.currentActiveH2Id = null; // Track active h2 for collapse/expand
-        this.boundHandleScroll = this.handleScroll.bind(this);
-        this.boundHandleResize = this.handleResize.bind(this);
-        this.ticking = false; // RAF throttle flag
-        this.cachedDetectionLine = 0; // Cache detection line position
-        this.initialScrollEndHandler = null; // One-time scrollend handler
-        this.initialScrollTimeout = null; // Fallback timeout for browsers without scrollend
-        this.initialized = false; // Track initialization state to prevent duplicate listeners
-        this.initialPagePath = null; // pathname+search when initialized; used to detect page changes
-    }
-
-    /**
-     * Initialize the scroll spy
-     */
-    init() {
-        // Guard against duplicate initialization
-        if (this.initialized) {
-            console.warn('TocScrollSpy already initialized. Call destroy() first if re-initialization is needed.');
-            return;
-        }
-
-        // Record the current page path (pathname+search, no hash) so we can detect
-        // when the user navigates away during enhanced navigation transitions.
-        // This prevents spurious replaceState calls from corrupting the new page's URL.
-        this.initialPagePath = window.location.pathname + window.location.search;
-
-        // Find all H2 and H3 headings in content with IDs
-        // H2 and H3 are always in the TOC, H4+ are not tracked
-        this.headings = Array.from(
-            this.contentElement.querySelectorAll('h2[id], h3[id]')
-        );
-
-        if (this.headings.length === 0) {
-            return;
-        }
-
-        // Build map of heading IDs to TOC links and cache heading elements
-        // Support both hash-only links (#id) and full URLs (http://...#id)
-        this.headings.forEach(heading => {
-            const id = heading.getAttribute('id');
-            const tocLink = this.tocElement.querySelector(`a[href="#${id}"], a[href$="#${id}"]`);
-            if (tocLink) {
-                this.tocLinks.set(id, tocLink);
-                this.headingElements.set(id, heading); // Cache heading element
-            }
-        });
-
-        // Cache h2 TOC items to avoid repeated querySelectorAll
-        this.h2Items = Array.from(this.tocElement.querySelectorAll('.toc-depth-0'));
-
-        // Calculate and cache detection line position
-        this.updateDetectionLine();
-
-        this.attachEventListeners();
-
-        // If a scroll restore is in progress (back/forward navigation), defer all initial
-        // heading detection. Running updateActiveHeading() here would see scrollY=0 (the
-        // browser-reset value before restoration), call setActive(null), strip the hash
-        // from the URL via replaceState, and potentially overwrite correct state.
-        // finishScrollRestore() in nav-helpers.js dispatches a synthetic scroll event once
-        // the position is restored, which triggers handleScroll() → updateActiveHeading()
-        // at the correct scrollY. No special handling needed here.
-        if (window.__scrollRestoring) {
-            // Skip initial detection entirely — let the post-restore synthetic scroll do it.
-        } else {
-            // Normal page load or forward navigation: detect immediately, then watch for
-            // scrollend in case the browser is still animating to a hash anchor.
-            this.updateActiveHeading();
-
-            // Handle case where browser is already scrolling when we initialize
-            // Set up one-time handler that auto-cancels when regular scroll/resize handlers trigger
-            if ('onscrollend' in window) {
-                this.initialScrollEndHandler = () => {
-                    if (window.location.pathname + window.location.search !== this.initialPagePath) {
-                        this.cleanupInitialScrollHandlers();
-                        return;
-                    }
-                    this.updateActiveHeading();
-                    this.cleanupInitialScrollHandlers();
-                };
-                window.addEventListener('scrollend', this.initialScrollEndHandler, { once: true });
-            } else {
-                // Fallback for browsers without scrollend support
-                this.initialScrollTimeout = setTimeout(() => {
-                    if (window.location.pathname + window.location.search !== this.initialPagePath) {
-                        this.cleanupInitialScrollHandlers();
-                        return;
-                    }
-                    this.updateActiveHeading();
-                    this.cleanupInitialScrollHandlers();
-                }, 200);
-            }
-        }
-
-        this.initialized = true;
-        if (typeof window.__e2eSignal === 'function') window.__e2eSignal('toc-initialized');
-    }
-
-    /**
-     * Attach event listeners for scroll tracking
-     * Called AFTER initial heading update to avoid race conditions
-     */
-    attachEventListeners() {
-        // Listen to scroll event with RAF throttling for real-time updates
-        window.addEventListener('scroll', this.boundHandleScroll, { passive: true });
-
-        // Recalculate detection line on window resize
-        window.addEventListener('resize', this.boundHandleResize, { passive: true });
-
-        // Note: No focusin handler - scroll event handles all collapse/expand updates
-        // This prevents DOM manipulation from interfering with browser navigation
-    }
-
-    /**
-     * Handle scroll events with requestAnimationFrame throttling
-     * This ensures updates happen at most once per frame (~60fps)
-     */
-    handleScroll() {
-        // During Blazor enhanced back/forward navigation the URL changes (popstate) before
-        // the DOM is swapped out. If this spy's scroll listener fires in that window it
-        // would call replaceState on the *new* page's URL, appending a stale heading hash
-        // (e.g. /all?types=videos → /all?types=videos#notes). That corrupted URL then
-        // causes Blazor's hash-scroll handler to fire on the browse page and reset scrollY
-        // to 0, defeating the scroll-position restoration and triggering an infinite-scroll
-        // cascade. Detecting the mismatch here and self-destructing breaks the chain.
-        if (this.initialPagePath !== null &&
-            window.location.pathname + window.location.search !== this.initialPagePath) {
-            this.destroy();
-            return;
-        }
-        // nav-helpers.js sets window.__scrollRestoring = true on popstate and clears it
-        // once scroll restoration completes. While set, skip all TOC updates to prevent
-        // replaceState calls that would append a stale heading hash to the restored URL.
-        if (window.__scrollRestoring) return;
-        this.cleanupInitialScrollHandlers(); // Cancel initial scroll handlers
-        if (!this.ticking) {
-            window.requestAnimationFrame(() => {
-                this.ticking = false;
-                // Guard: destroy() may have been called while this RAF was pending
-                // (e.g. user clicked a link before this frame fired). Without this check,
-                // updateActiveHeading() would run on the old DOM after the URL changed and
-                // call setActive(null) → replaceState, stripping the hash from the new page's URL.
-                if (!this.initialized) return;
-                this.updateActiveHeading();
-            });
-            this.ticking = true;
-        }
-    }
-
-    /**
-     * Handle window resize - recalculate detection line position
-     */
-    handleResize() {
-        this.cleanupInitialScrollHandlers(); // Cancel initial scroll handlers
-        this.updateDetectionLine();
-    }
-
-    /**
-     * Calculate and cache detection line position
-     * Uses hardcoded constants that MUST match CSS scroll-margin-top value
-     */
-    updateDetectionLine() {
-        const contentViewportHeight = window.innerHeight - STICKY_HEADER_OFFSET;
-        this.cachedDetectionLine = STICKY_HEADER_OFFSET + (contentViewportHeight * DETECTION_LINE_PERCENT);
-    }
-
-    /**
-     * Update which heading should be active based on scroll position
-     */
-    updateActiveHeading() {
-        // Use cached detection line position
-        const detectionLineFromTop = this.cachedDetectionLine;
-        const tolerance = 5; // Allow 5px tolerance
-
-        // Bottom-of-page detection: when scrolled to the very bottom, activate the last heading.
-        // This ensures all sections are reachable via scrolling - a standard scroll spy UX pattern.
-        const scrollBottom = window.innerHeight + window.scrollY;
-        const pageHeight = document.documentElement.scrollHeight;
-        if (pageHeight - scrollBottom < 50 && this.headings.length > 0) {
-            const lastHeading = this.headings[this.headings.length - 1];
-            const lastId = lastHeading.getAttribute('id');
-            if (lastId && this.tocLinks.has(lastId)) {
-                this.setActive(lastId);
-                return;
-            }
-        }
-
-        // Find the heading CLOSEST to our detection line (30% from top of content area)
-        // Pick the heading that's above the line and closest to it
-        let targetId = null;
-        let closestDistance = Infinity;
-
-        // CRITICAL: Only check headings that are actually in the TOC (have a link in tocLinks)
-        // This prevents H5/H6 or other unlisted headings from triggering highlights
-        for (const [headingId, heading] of this.headingElements) {
-            const rect = heading.getBoundingClientRect();
-            const headingTopFromViewport = rect.top;
-
-            // Only consider headings that are at or above the detection line (with tolerance)
-            if (headingTopFromViewport <= detectionLineFromTop + tolerance) {
-                const distance = Math.abs(detectionLineFromTop - headingTopFromViewport);
-
-                // Pick the heading closest to the detection line (smallest distance)
-                if (distance < closestDistance) {
-                    closestDistance = distance;
-                    targetId = headingId;
-                }
-            }
-        }
-
-        // If no heading is above the detection line (e.g., page just loaded at top),
-        // don't highlight anything — wait until user scrolls
-        if (!targetId) {
-            // Clear any existing active state
-            if (this.currentActiveId !== null) {
-                this.setActive(null);
-            }
-            return;
-        }
-
-        if (targetId) {
-            this.setActive(targetId);
-        }
-    }
-
-    /**
-     * Set a TOC link as active
-     */
-    setActive(headingId, skipCollapseUpdate = false) {
-        // Don't update if already active
-        if (this.currentActiveId === headingId) {
-            return;
-        }
-
-        // Remove active from current TOC link and heading
-        if (this.currentActiveId !== null) {
-            const currentLink = this.tocLinks.get(this.currentActiveId);
-            if (currentLink) {
-                currentLink.classList.remove('active');
-            }
-
-            const currentHeading = this.headingElements.get(this.currentActiveId);
-            if (currentHeading) {
-                currentHeading.classList.remove('toc-active-heading');
-            }
-        }
-
-        // If clearing active state (null), remove hash from URL and reset
-        if (headingId === null) {
-            this.currentActiveId = null;
-            // Remove hash from URL when no heading is above the detection line
-            if (window.location.hash) {
-                const cleanUrl = `${window.location.pathname}${window.location.search}`;
-                history.replaceState(null, '', cleanUrl);
-            }
-            return;
-        }
-
-        // Use cached heading element
-        const newHeading = this.headingElements.get(headingId);
-        if (!newHeading) {
-            return;
-        }
-
-        // Add active to new TOC link and heading
-        const newLink = this.tocLinks.get(headingId);
-        if (newLink) {
-            newLink.classList.add('active');
-            this.currentActiveId = headingId;
-
-            // Update URL hash without polluting browser history
-            // replaceState keeps back button clean (only TOC clicks create history entries)
-            const newUrl = `${window.location.pathname}${window.location.search}#${headingId}`;
-            history.replaceState(null, '', newUrl);
-
-            // Add active class to current heading
-            newHeading.classList.add('toc-active-heading');
-
-            if (typeof window.__e2eSignal === 'function') window.__e2eSignal('toc-active-updated');
-
-            // Update collapse/expand state based on which h2 section we're in (unless skipped)
-            if (!skipCollapseUpdate) {
-                this.updateCollapseState(headingId);
-            }
-        }
-    }
-
-    /**
-     * Update which h2 section is expanded based on active heading
-     */
-    updateCollapseState(activeHeadingId) {
-        // Use cached heading element
-        const activeHeading = this.headingElements.get(activeHeadingId);
-        if (!activeHeading) return;
-
-        // Get the heading level
-        const headingLevel = parseInt(activeHeading.tagName.substring(1)); // h2 -> 2, h3 -> 3
-
-        let targetH2Id = activeHeadingId;
-
-        // If this is an h3, find the parent h2
-        if (headingLevel === 3) {
-            // Walk backwards through headings to find the parent h2
-            const activeIndex = this.headings.indexOf(activeHeading);
-            if (activeIndex === -1) return; // Heading not in our list, abort
-
-            for (let i = activeIndex - 1; i >= 0; i--) {
-                const heading = this.headings[i];
-                const level = parseInt(heading.tagName.substring(1));
-                if (level === 2) {
-                    targetH2Id = heading.getAttribute('id');
-                    break;
-                }
-            }
-        }
-
-        // Verify the target H2 has a TOC link before proceeding
-        // If not, we can't expand any section, so just return early
-        const activeTocLink = this.tocLinks.get(targetH2Id);
-        if (!activeTocLink) return;
-
-        // Only update if we've changed h2 sections
-        if (this.currentActiveH2Id === targetH2Id) {
-            return;
-        }
-
-        // Find the new item to expand
-        const activeItem = activeTocLink.closest('.toc-depth-0');
-
-        // FIRST: Expand the new section
-        if (activeItem) {
-            activeItem.classList.add('expanded');
-        }
-
-        // SECOND: Collapse all OTHER sections (exclude the newly expanded one)
-        this.h2Items.forEach(item => {
-            if (item !== activeItem) {
-                item.classList.remove('expanded');
-            }
-        });
-
-        this.currentActiveH2Id = targetH2Id;
-    }
-
-    /**
-     * Clean up initial scroll handlers
-     * Called automatically when regular scroll/resize handlers trigger
-     */
-    cleanupInitialScrollHandlers() {
-        if (this.initialScrollEndHandler) {
-            window.removeEventListener('scrollend', this.initialScrollEndHandler);
-            this.initialScrollEndHandler = null;
-        }
-        if (this.initialScrollTimeout) {
-            clearTimeout(this.initialScrollTimeout);
-            this.initialScrollTimeout = null;
-        }
-    }
-
-    /**
-     * Clean up event listeners
-     */
-    destroy() {
-        this.cleanupInitialScrollHandlers();
-        window.removeEventListener('scroll', this.boundHandleScroll);
-        window.removeEventListener('resize', this.boundHandleResize);
-
-        this.initialized = false;
-    }
-}
-
-/**
- * Initialize all TOC scroll spies on the page
- */
-export function initTocScrollSpy() {
-    const TABLET_BREAKPOINT = 1292;
-    const isMobile = window.innerWidth <= TABLET_BREAKPOINT;
-    const tocElements = document.querySelectorAll('[data-toc-scroll-spy]');
-
-    tocElements.forEach(tocElement => {
-        const contentSelector = tocElement.getAttribute('data-content-selector');
-        if (!contentSelector) {
-            console.warn('TOC element missing data-content-selector attribute');
-            return;
-        }
-
-        const contentElement = document.querySelector(contentSelector);
-        if (!contentElement) {
-            console.warn(`Content element not found: ${contentSelector}`);
-            return;
-        }
-
-        // Clean up existing instance if it exists
-        if (tocElement._tocScrollSpy) {
-            tocElement._tocScrollSpy.destroy();
-            tocElement._tocScrollSpy = null;
-        }
-
-        // Clean up existing mobile click handlers
-        if (tocElement._mobileClickHandler) {
-            tocElement.removeEventListener('click', tocElement._mobileClickHandler);
-            tocElement._mobileClickHandler = null;
-        }
-
-        if (isMobile) {
-            // Mobile: no scroll spy, just tap-to-toggle on H2 items
-            tocElement.classList.add('toc-mobile-mode');
-
-            const clickHandler = (e) => {
-                const tocLink = e.target.closest('.toc-depth-0 > .toc-link');
-                if (!tocLink) return;
-
-                const tocItem = tocLink.closest('.toc-depth-0');
-                if (!tocItem) return;
-
-                // Only toggle if this H2 has sub-items
-                const sublist = tocItem.querySelector('.toc-sublist');
-                if (!sublist) return;
-
-                e.preventDefault();
-                tocItem.classList.toggle('expanded');
-            };
-
-            tocElement.addEventListener('click', clickHandler);
-            tocElement._mobileClickHandler = clickHandler;
-        } else {
-            // Desktop: full scroll spy
-            tocElement.classList.remove('toc-mobile-mode');
-
-            const scrollSpy = new TocScrollSpy(tocElement, contentElement);
-            scrollSpy.init();
-
-            // Store instance for cleanup if needed
-            tocElement._tocScrollSpy = scrollSpy;
-        }
-    });
-}
-
-/**
- * Clean up all TOC scroll spy instances
- */
-export function cleanupAllTocScrollSpies() {
-    const tocElements = document.querySelectorAll('[data-toc-scroll-spy]');
-
-    tocElements.forEach(tocElement => {
-        if (tocElement._tocScrollSpy) {
-            tocElement._tocScrollSpy.destroy();
-            tocElement._tocScrollSpy = null;
-        }
-        if (tocElement._mobileClickHandler) {
-            tocElement.removeEventListener('click', tocElement._mobileClickHandler);
-            tocElement._mobileClickHandler = null;
-        }
-        tocElement.classList.remove('toc-mobile-mode');
-    });
-}
