+++ tests/TechHub.Web.Tests/Middleware/UrlNormalizationMiddlewareTests.cs
@@ -710,8 +710,8 @@ public class UrlNormalizationMiddlewareTests
     [Theory]
     [InlineData("/css/article.css")]
     [InlineData("/css/sidebar.css")]
-    [InlineData("/js/nav-helpers.js")]
-    [InlineData("/js/infinite-scroll.js")]
+    [InlineData("/js/scroll-manager.js")]
+    [InlineData("/js/page-scripts.js")]
     [InlineData("/images/logo.png")]
     [InlineData("/images/section-backgrounds/ai.jxl")]
     [InlineData("/images/section-backgrounds/github-copilot.webp")]
